feat(GameManager): Minigame Hand generation
Now GenerateHands (previously named EndCalibration) is also called when a new scene is loaded.  When the scene is a menu, generates cursors, sets their parent to Canvas and then their position to Vector3.zero. When on a minigame, spawns its respective hands in the position of an object called HandSpawner